requirements.txt
rich
requests
cassandra-driver
streamlit